lpc: remove hardcoded function name checks from builtin_refs codegen
- Extract actual Rust function signatures from parsed source code
- Use format_rust_function_signature() to get real types
- Remove all hardcoded name checks (lpfx_hash_, lpfx_simplex, etc.)
- Make rust_signature non-optional - fail if signature can't be parsed
- Generated builtin_refs.rs now uses correct f32 types for _f32 functions

+++ lp-glsl/apps/lp-builtins-app/src/builtin_refs.rs
@@ -60,11 +60,11 @@ pub fn ensure_builtins_referenced() {
         let __lpfx_hash_1_fn: extern "C" fn(u32, u32) -> u32 = __lpfx_hash_1;
         let __lpfx_hash_2_fn: extern "C" fn(u32, u32, u32) -> u32 = __lpfx_hash_2;
         let __lpfx_hash_3_fn: extern "C" fn(u32, u32, u32, u32) -> u32 = __lpfx_hash_3;
-        let __lpfx_simplex1_f32_fn: extern "C" fn(i32, u32) -> i32 = __lpfx_simplex1_f32;
+        let __lpfx_simplex1_f32_fn: extern "C" fn(f32, u32) -> f32 = __lpfx_simplex1_f32;
         let __lpfx_simplex1_q32_fn: extern "C" fn(i32, u32) -> i32 = __lpfx_simplex1_q32;
-        let __lpfx_simplex2_f32_fn: extern "C" fn(i32, i32, u32) -> i32 = __lpfx_simplex2_f32;
+        let __lpfx_simplex2_f32_fn: extern "C" fn(f32, f32, u32) -> f32 = __lpfx_simplex2_f32;
         let __lpfx_simplex2_q32_fn: extern "C" fn(i32, i32, u32) -> i32 = __lpfx_simplex2_q32;
-        let __lpfx_simplex3_f32_fn: extern "C" fn(i32, i32, i32, u32) -> i32 = __lpfx_simplex3_f32;
+        let __lpfx_simplex3_f32_fn: extern "C" fn(f32, f32, f32, u32) -> f32 = __lpfx_simplex3_f32;
         let __lpfx_simplex3_q32_fn: extern "C" fn(i32, i32, i32, u32) -> i32 = __lpfx_simplex3_q32;
 
         // Force these to be included by using them in a way that can't be optimized away
